requirement.txt
imgkit==1.0.2
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.5
Pillow==8.0.1
python-barcode==0.13.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
wkhtmltopdf==0.2
xlrd==2.0.1
